fixed search query bug + organized components
