+++ src/App.js
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import Form from './Form';
-import DiscoveryLanding from './DiscoveryLanding';
-import MovieDetails from './MovieDetails';
-import SearchMovie from './SearchMovie';
+import Form from './Components/Form';
+import DiscoveryLanding from './Components/DiscoveryLanding';
+import MovieDetails from './Components/MovieDetails';
+import SearchMovie from './Components/SearchMovie';
 
 function App() {
 
@@ -22,6 +22,6 @@ function App() {
 }
 
 export default App;
-
+// Anything after ":" is technically a prop, which we have access to on SearchMovies.js
 // Want to pass parameter into component, need to use render={} instead of component={}
 // render={() => <SearchMovie searchQuery={searchQuery} /> }
