+++ src/Components/Firebase.js
@@ -1,0 +1,10 @@
+const firebaseConfig = {
+    apiKey: "AIzaSyDefbwUJmMQ00xMEYJKwQgK2qWbD6HnrRI",
+    authDomain: "movie-trio.firebaseapp.com",
+    projectId: "movie-trio",
+    storageBucket: "movie-trio.appspot.com",
+    messagingSenderId: "709082618933",
+    appId: "1:709082618933:web:f0e32576975eccb542e57c"
+};
+// Initialize Firebase
+firebase.initializeApp(firebaseConfig);
